+++ Assets/Prefabs/Hex Grid Chunk.prefab
@@ -45,6 +45,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 01d46b3c49dde4747b289b3ca5823df1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  terrain: {fileID: 8622655605179442443}
 --- !u!1001 &3952967181139723918
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -180,7 +181,7 @@ PrefabInstance:
     - target: {fileID: 1289660638312453624, guid: 069a7d00570a84ccca74f2f5b5c7048f,
         type: 3}
       propertyPath: m_Name
-      value: Hex Mesh
+      value: Terrain
       objectReference: {fileID: 0}
     - target: {fileID: 1289660638312453627, guid: 069a7d00570a84ccca74f2f5b5c7048f,
         type: 3}
@@ -237,6 +238,16 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 1289660638312453626, guid: 069a7d00570a84ccca74f2f5b5c7048f,
+        type: 3}
+      propertyPath: useCollider
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1289660638312453626, guid: 069a7d00570a84ccca74f2f5b5c7048f,
+        type: 3}
+      propertyPath: useColors
+      value: 1
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 069a7d00570a84ccca74f2f5b5c7048f, type: 3}
 --- !u!4 &8622655605179442442 stripped
@@ -245,3 +256,15 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 7371293155891261681}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &8622655605179442443 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 1289660638312453626, guid: 069a7d00570a84ccca74f2f5b5c7048f,
+    type: 3}
+  m_PrefabInstance: {fileID: 7371293155891261681}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2ebf814afc44a463bad4553d7e077143, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
